fix(engine): apply clippy lint fixes across 18 engine modules
Systematic clippy cleanup touching attestation_handshake,
conformance_vector_gen, containment_executor,
controller_interference_guard, declassification_pipeline,
demotion_rollback, evidence_contract, evidence_emission,
extension_host_lifecycle, gc_pause, guardplane_calibration,
hostcall_telemetry, interleaving_explorer, lease_tracker, mmr_proof,
obligation_integration, obligation_leak_policy,
succinct_witness_compiler, and local_semantic_atlas_integration tests.

Changes include:
- `vec![...]` → `[...]` for array/slice literals that don't need Vec
- `.clone()` → `*` (dereference) for Copy types
- `assert!(x.len() >= 1)` → `assert!(!x.is_empty())`
- `assert_eq!(val, true/false)` → `assert!(val)` / `assert!(!val)`
- `assert!(!(a == b))` → `assert!(a != b)`
- `for (_id, v) in &map` → `for v in map.values()`
- Identity cast removal (`i as u64` → `i` where i is already u64)
- `bag_peaks(&[h.clone()])` → `bag_peaks(std::slice::from_ref(&h))`

No behavioral changes — all existing tests continue to pass.

+++ crates/franken-engine/src/gc_pause.rs
@@ -922,7 +922,7 @@ mod tests {
         let run = || {
             let mut tracker = PauseTracker::default();
             for i in 1..=50 {
-                tracker.record(&make_event(i, "ext-a", i * 100, i as u64, i * 10));
+                tracker.record(&make_event(i, "ext-a", i * 100, i, i * 10));
             }
             tracker.global_percentiles()
         };
